+++ src/utils/conftree.cpp
@@ -68,7 +68,7 @@ void ConfSimple::parseinput(istream &input)
 	}
 
         {
-            int ll = strlen(cline);
+            size_t ll = strlen(cline);
             while (ll > 0 && (cline[ll-1] == '\n' || cline[ll-1] == '\r')) {
                 cline[ll-1] = 0;
                 ll--;
